refactor: refactor Redux Store
