+++ src/components/Layout/Layout.tsx
@@ -1,7 +1,7 @@
 import SearchBar from '../SearchBar/SearchBar';
 import List from '../List/List';
 import { useRouter } from 'next/router';
-import { useGetCharactersQuery } from '@/store/query/api';
+import { useGetCharactersQuery } from '@/store/query/disneyApi';
 
 type PropsType = {
   children?: React.ReactNode;
@@ -13,8 +13,8 @@ const Layout: React.FC<PropsType> = ({ children }) => {
 
   const { data } = useGetCharactersQuery({
     search: (search as string) || '',
-    pageSize: +(limit as string) || 10,
-    pageNumber: +(page as string) || 1,
+    pageSize: (limit as string) || '10',
+    pageNumber: (page as string) || '1',
   });
 
   return (
